requirements.txt
flask==2.0.1
python-dotenv==0.19.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-storage-blob==12.9.0
requests==2.26.0
Pillow==8.3.2
